# api/docs/swagger.yaml
basePath: /
definitions:
...
  email.EmailResponse:
    properties:
      message:
        type: string
      message_id:
        type: string
      sent_to:
        items:
          type: string
        type: array
      success:
        type: boolean
    type: object
  email.SendEmailRequest:
    properties:
      bcc:
        items:
          type: string
        type: array
      body:
        minLength: 1
        type: string
      cc:
        items:
          type: string
        type: array
      is_html:
        type: boolean
      subject:
        maxLength: 500
        minLength: 1
        type: string
      to:
        items:
          type: string
        minItems: 1
        type: array
    required:
    - body
    - subject
    - to
    type: object
  email.SendTemplateEmailRequest:
    properties:
      bcc:
        items:
          type: string
        type: array
      cc:
        items:
          type: string
        type: array
      subject:
        maxLength: 500
        minLength: 1
        type: string
      template_data:
        additionalProperties: true
        type: object
      template_name:
        enum:
        - default
        - welcome
        - notification
        type: string
      to:
        items:
          type: string
        minItems: 1
        type: array
    required:
    - subject
    - template_name
    - to
    type: object
  errors.ErrorInfo:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
      path:
        type: string
      request_id:
        type: string
      retry_after:
        type: integer
      timestamp:
        type: string
    type: object
  errors.Links:
    properties:
      first:
        type: string
      last:
        type: string
      next:
        type: string
      prev:
        type: string
      self:
        type: string
    type: object
  errors.Meta:
    properties:
      links:
        $ref: '#/definitions/errors.Links'
      page:
        type: integer
      per_page:
        type: integer
      request_id:
        type: string
      timestamp:
        type: string
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  errors.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/errors.ErrorInfo'
      meta:
        $ref: '#/definitions/errors.Meta'
      success:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A production-ready REST API boilerplate in Go with JWT authentication
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: TRIIIO API Boilerplate
  version: "1.0"
paths:
...
  /api/v1/emails/send:
    post:
      consumes:
      - application/json
      description: Send a simple email to one or more recipients
      parameters:
      - description: Email data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/email.SendEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/errors.Response'
            - properties:
                data:
                  $ref: '#/definitions/email.EmailResponse'
                success:
                  type: boolean
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/errors.Response'
            - properties:
                error:
                  $ref: '#/definitions/errors.ErrorInfo'
                success:
                  type: boolean
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/errors.Response'
            - properties:
                error:
                  $ref: '#/definitions/errors.ErrorInfo'
                success:
                  type: boolean
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/errors.Response'
            - properties:
                error:
                  $ref: '#/definitions/errors.ErrorInfo'
                success:
                  type: boolean
              type: object
      security:
      - BearerAuth: []
      summary: Send email
      tags:
      - emails
  /api/v1/emails/send-template:
    post:
      consumes:
      - application/json
      description: Send an email using a predefined HTML template
      parameters:
      - description: Template email data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/email.SendTemplateEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/errors.Response'
            - properties:
                data:
                  $ref: '#/definitions/email.EmailResponse'
                success:
                  type: boolean
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/errors.Response'
            - properties:
                error:
                  $ref: '#/definitions/errors.ErrorInfo'
                success:
                  type: boolean
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/errors.Response'
            - properties:
                error:
                  $ref: '#/definitions/errors.ErrorInfo'
                success:
                  type: boolean
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/errors.Response'
            - properties:
                error:
                  $ref: '#/definitions/errors.ErrorInfo'
                success:
                  type: boolean
              type: object
      security:
      - BearerAuth: []
      summary: Send template email
      tags:
      - emails
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"